src/nonogram.py
'''
Created on 2014/09/14

@author: alchu
'''

class Nonogram(object):
    def __init__(self):
        self._left = []
        self._top = []
        self._field = []
    
    def set_left(self, ary):
        self._left = ary
        
    def set_top(self, ary):
        
        self._top =ary
        
    def row(self):
        return len(self._left)
    
    def column(self):
        return len(self._top)
    
    def solve(self):
        '''stub'''
        self._field = [0,0]
    
if __name__ == '__main__':
    pass